fix(fail): move DetailedError and exit codes to internal/fail to fix CONSTITUTION violation
Extract the shared error types (DetailedError, exit code constants, WriteJSON methods)
out of cmd/gcx/fail into a new internal/fail package so that internal/ packages can
import them without violating the cmd/→internal/ one-way dependency rule.

cmd/gcx/fail re-exports the type and constants via a type alias and const re-exports
so all existing cmd/ callers are unaffected. Tests for the type's own behaviour are
co-located with the type in internal/fail; convert.go-specific tests remain in cmd/gcx/fail.

Fixes: internal/providers/fleet importing cmd/gcx/fail (CONSTITUTION § Dependency Rules).
